Fix: Move historical node registry to utils module
