+++ utils/__init__.py
@@ -1,8 +1,8 @@
+from utils.bls import get_aggregated_public_key
+from utils.nodes_registry import NodesRegistryClient
 from utils.nodes_registry import fetch_historical_nodes_registry_data, get_nodes_registry_last_tag
 from utils.sub_graph import get_eigen_network_info, fetch_eigen_layer_last_block_number
 from utils.utils import get_file_content, validate_env_variables
-from utils.bls import get_aggregated_public_key
-from utils.historical_nodes_registry.client import NodesRegistryClient
 
 __all__ = ['NodesRegistryClient',
            'get_file_content',
