--- utils/historical_nodes_registry/__init__.py
@@ -1,8 +1,0 @@
-# from utils.historical_nodes_registry.client import NodesRegistryClient
-# from utils.historical_nodes_registry.schema import NodeInfo, SnapShotType
-#
-# __all__ = [
-#     'NodesRegistryClient',
-#     'NodeInfo',
-#     'SnapShotType'
-# ]
